Add files via upload
Here is my Blender character rig. 

IK rig and parenting works fairly well with the arms but not so much for the legs and I can't figure the reason. I've tried to undo all that I've done to the legs so that it's a fresh start. 
Hoping someone can help me on this rigging project as it's my first and I'm still learning. 

Thanks for all the help in advance!
-Matt